Add logg_load(char *name)
This function searches the environment LOGGER_PATH variable
for directories to load a module ch_<name>.so, where a channel
description is defined.  It dynamically loads the module and
executes the _init() function, which is supposed to register
the struct logg_chann_ops structure with name "<name>" via
the logg_register_channops() function.

Signed-off-by: Luis Colorado <luiscoloradosistemas@gmail.com>